[P2300] minor changes
- refactor: change the meta types location to meta header
- Ref - https://github.com/STEllAR-GROUP/hpx/pull/5990/#discussion_r993437346
- https://github.com/STEllAR-GROUP/hpx/pull/5990/#discussion_r993437947

- change return type to constexpr bool for
  sender_awaitable_base::await_ready
  Ref - https://github.com/STEllAR-GROUP/hpx/pull/5990/#discussion_r993440986
