+++ libs/core/type_support/include/hpx/type_support/meta.hpp
@@ -453,4 +453,29 @@ namespace hpx { namespace meta {
         using apply = invoke<compose_args<Continuation>,
             if_<std::is_same<Ts, Old>, pack<>, pack<Ts>>...>;
     };
+
+    template <typename A, typename... As>
+    struct front
+    {
+        using type = A;
+    };
+
+    template <template <typename...> typename Fn>
+    struct compose_template_func
+    {
+        template <typename... Args>
+        using apply = Fn<Args...>;
+    };
+
+    template <typename... As>
+    using single_t =
+        std::enable_if_t<sizeof...(As) == 1, meta::type<front<As...>>>;
+
+    template <typename Ty>
+    struct single_or
+    {
+        template <typename... As>
+        using apply =
+            std::enable_if_t<sizeof...(As) <= 1, meta::type<front<As..., Ty>>>;
+    };
 }}    // namespace hpx::meta
